split b903c00:src/event_handler/mod.rs (no behavior changes)
